Now assembles a ret instruction.
Changed the test and assembler environments now copy in the right build.bat for ease of compiling.

win64_assembler.h now has my QoL types and some helper functions.

win64_test.cpp initializes the assembler and gives it the chunks of memory it's looking for.

The assembler now requires initialization so I can use a helper function to add reserved keywords to a list for now.

The assembler now correctly compiles the return statement and sort of identifies a number.