Gearing up to implement lea opcode and finished string initialization for now.
Going to use AlignSize to calculate RIP-relative addresses. Now pass in PAGE because that's the section alignment to help with that calculation.

Started to implement lea opcode.

Started to implement a dynamic array of code patches. Which gets cleared when all the code has been assembled.

When there's been a string added to the resource section and a new variable has been reserved on the same line, the string info is stored in the new variable when the line ends.