Import data tables for functions now added to .exe
Put a toggle in the analyze function because, for now, for debug purposes I need to analyze the .exe file the assembler puts out.

Changed what I wanted the code to look like.

Added imported functions as their own type and reserved the two variables that are going to be used.

Added a helper function to put strings into an array.

Now reserve all the space at the top of the resource section for the null terminated tables that Windows uses. The assemble function even fills them in. I might regret that later.