Can now read from the stack
Now properly fills in the operand when a variable gets called again after its creation.

The op code is what determines whether something is being written to memory or read from memory, so mov now properly picks the op code to use and identifies the source and the destination. Successfully reads from the stack now. Has a new test to verify that.